fix: postgresqlLoader
- Change file name
- Load config from config module
